adjust the file dirs
