+++ src/query/storages/fuse/src/operations/common/mod.rs
@@ -12,8 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod abort_operation;
+mod mutation_accumulator;
+mod mutation_log;
+mod processors;
 mod snapshot_generator;
 
-pub use snapshot_generator::AppendGenerator;
-pub use snapshot_generator::MutationGenerator;
-pub use snapshot_generator::SnapshotGenerator;
+pub use abort_operation::AbortOperation;
+pub use mutation_accumulator::MutationAccumulator;
+pub use mutation_log::*;
+pub use processors::*;
+pub use snapshot_generator::*;
