+++ src/query/storages/fuse/src/operations/common/mutation_accumulator.rs
@@ -34,12 +34,12 @@ use tracing::info;
 use crate::io::SegmentsIO;
 use crate::io::SerializedSegment;
 use crate::io::TableMetaLocationGenerator;
-use crate::operations::merge_into::mutation_meta::AppendOperationLogEntry;
-use crate::operations::merge_into::mutation_meta::CommitMeta;
-use crate::operations::merge_into::mutation_meta::MutationLogEntry;
-use crate::operations::merge_into::mutation_meta::Replacement;
-use crate::operations::merge_into::mutation_meta::ReplacementLogEntry;
-use crate::operations::mutation::AbortOperation;
+use crate::operations::common::AbortOperation;
+use crate::operations::common::AppendOperationLogEntry;
+use crate::operations::common::CommitMeta;
+use crate::operations::common::MutationLogEntry;
+use crate::operations::common::Replacement;
+use crate::operations::common::ReplacementLogEntry;
 use crate::operations::mutation::BlockIndex;
 use crate::operations::mutation::SegmentIndex;
 use crate::statistics::reducers::deduct_statistics_mut;
