+++ src/query/storages/fuse/src/operations/common/processors/fill_internal_columns.rs
