+++ src/query/storages/fuse/src/operations/common/processors/sink_commit.rs
@@ -36,9 +36,9 @@ use table_lock::TableLockHeartbeat;
 use crate::io::TableMetaLocationGenerator;
 use crate::metrics::metrics_inc_commit_aborts;
 use crate::metrics::metrics_inc_commit_mutation_success;
-use crate::operations::merge_into::mutation_meta::CommitMeta;
-use crate::operations::mutation::AbortOperation;
-use crate::operations::SnapshotGenerator;
+use crate::operations::common::AbortOperation;
+use crate::operations::common::CommitMeta;
+use crate::operations::common::SnapshotGenerator;
 use crate::pipelines::processors::port::InputPort;
 use crate::pipelines::processors::processor::Event;
 use crate::pipelines::processors::processor::ProcessorPtr;
