+++ src/query/storages/fuse/src/operations/mutation/compact/block_compact_mutator.rs
@@ -33,7 +33,7 @@ use storages_common_table_meta::meta::Statistics;
 use tracing::info;
 
 use crate::io::SegmentsIO;
-use crate::operations::merge_into::mutation_meta::BlockMetaIndex;
+use crate::operations::common::BlockMetaIndex;
 use crate::operations::mutation::CompactPartInfo;
 use crate::operations::CompactOptions;
 use crate::statistics::reducers::deduct_statistics_mut;
