+++ src/query/storages/fuse/src/operations/mutation/compact/compact_aggregator.rs
@@ -35,9 +35,9 @@ use tracing::info;
 use crate::io::SegmentsIO;
 use crate::io::SerializedSegment;
 use crate::io::TableMetaLocationGenerator;
-use crate::operations::merge_into::mutation_meta::CommitMeta;
+use crate::operations::common::AbortOperation;
+use crate::operations::common::CommitMeta;
 use crate::operations::mutation::compact::CompactSourceMeta;
-use crate::operations::mutation::AbortOperation;
 use crate::operations::mutation::BlockCompactMutator;
 use crate::statistics::reducers::merge_statistics_mut;
 use crate::statistics::reducers::reduce_block_metas;
